+++ common/base_cpp/smart_output.h
@@ -30,8 +30,8 @@ public:
    virtual ~SmartTableOutput ();
 
    virtual void write (const void *data, int size);
-   virtual void seek  (int offset, int from);
-   virtual int  tell  ();
+   virtual void seek  (long long offset, int from);
+   virtual long long tell  ();
    virtual void flush ();
 
    void setLineFormat (const char *line_format);
