+++ common/gzip/gzip_output.cpp
@@ -113,12 +113,12 @@ int GZipOutput::_deflate (int flush)
    return rc;
 }
 
-int GZipOutput::tell ()
+long long GZipOutput::tell ()
 {
    return _total_written;
 }
 
-void GZipOutput::seek (int offset, int from)
+void GZipOutput::seek (long long offset, int from)
 {
    throw Error("not imlemented");
 }
