+++ common/gzip/gzip_output.h
@@ -31,8 +31,8 @@ public:
    virtual ~GZipOutput ();
 
    virtual void write (const void *data, int size);
-   virtual void seek  (int offset, int from);
-   virtual int  tell  ();
+   virtual void seek  (long long offset, int from);
+   virtual long long tell  ();
    virtual void flush ();
 
    DECL_ERROR;
